Этап17

+++ src/main/java/ApplicationTask/TaskDaoJdbcTemplate.java
@@ -3,11 +3,13 @@ package ApplicationTask;
 import ApplicationTask.Task;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.SystemEnvironmentPropertySource;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.jdbc.core.JdbcTemplate;
 
 import java.sql.Array;
 import java.util.*;
@@ -15,29 +17,50 @@ import java.util.*;
 
 @Component
 
-public  class TaskDaoJdbcTemplate {
+public  class TaskDaoJdbcTemplate implements TaskDao {
     @Autowired
     private JdbcTemplate jdbcTemplate;
+    public String s = "";
 
-
-    public  String s;
-
-
-
-    public  String  bd_load()
-    {
-
-        jdbcTemplate.query("SELECT task_text FROM TASK",(rs, rowNum) ->
-                new Task(rs.getString("task_text"))).
-                forEach(task -> s=(task.toString()));
-        return  s;
+    public Task add(Task task) {
+        return new Task();
     }
 
-    public void bd_save(String mes)
-    {
-        jdbcTemplate.execute("drop table TASK");
-        jdbcTemplate.execute("CREATE TABLE TASK (task_text TEXT NOT NULL);");
-        jdbcTemplate.execute("INSERT INTO TASK (task_text) VALUES ('"+ mes+"')");
+    ;
 
+    public void delete(Integer id) {
     }
+
+    ;
+
+    public String load() {
+        List<Task> load_task =
+                jdbcTemplate.query("SELECT task_text FROM TASK", (rs, rowNum) ->
+                        //  new Task(rs.getString("task_text"))).
+                        //  forEach(task -> ));
+
+
+                {
+                    Task task = new Task();
+                    task.setText(rs.getString("task_text"));
+                    return task;
+                });
+        String stroka = "";
+        for (Task name : load_task) {
+            stroka += name.toString();
+
+        }
+        return stroka;
+    }
+
+
+    public void save(Task task) {
+        if(task.getId()==1) {
+            jdbcTemplate.execute("drop table TASK");
+            jdbcTemplate.execute("CREATE TABLE TASK (id INT ,task_text TEXT NOT NULL);");
+        }
+        jdbcTemplate.execute("INSERT INTO TASK  VALUES ("+task.getId()+",'"+ task.getText() + "')");
+
+}
+
 }
